drm_hwcomposer: Add platform backend for Qcom Gralloc
This adds support for the Qualcomm version of Gralloc.

The gralloc handle is not the same as the common libdrm handle,
so we do need a separate backend for now.

A imported buffer caching mechanism is applied here to prevent
the same buffer to be import/unimport for each composition cycle.
The buffer unimporting from the DRM will be delayed until the
cache reaches a certain size (16 by default).

Change-Id: I817d1d80b56ed6713289166d9e8167e4d7e45a30
Signed-off-by: Jim Wang <quic_jimwan@quicinc.com>