drm_hwcomposer: Remove HWC2 SetActiveConfigWithConstraints
Move the HWC2 implementation into the hwc2 frontend, and update the
implementation to call QueueConfig.

This is simpler than the hwc3 implementation, but should be fine since
QueueConfig is consistent with how HWC2 GetActiveConfig was implemented.
Specifically, by queuing up the config change to take effect along with
the next composition.

Change-Id: I448728bd93b5a35badba179b05e2c8865d62f123
Signed-off-by: Drew Davenport <ddavenport@google.com>