drm_hwcomposer: Remove GetDisplayConnectionType and GetDisplayType
GetDisplayType is not used for hwc3, so no frontend work is
required to remove it.

A display can be physical or virtual. A physical display can be internal
or external. Create an enum to cover these three cases (virtual,
internal, external), and add a function GetDisplayType.

Implement related functionality in the respective frontends in terms of
this new function.

Change-Id: I72b0b66d22bc91a7e6d0450469f4c140b2cb91c4
Signed-off-by: Drew Davenport <ddavenport@google.com>