drm_hwcomposer: Handle empty and cursor-only composition
It's possible for the client may validate/present a composition without
any layers. Early out in this case. This avoids a spurious
failed_kms_validate_ when CreateComposition returns an error in this
case.

Reduce the error in HwcDisplay::CreateComposition to a warning, since
the rest of the code should work even if an empty DrmKmsPlan is created.

Additionally, the error in HwcDisplay::CreateComposition did not
handle the case where the only layer is the cursor. This case is now
handled by checking for an empty composition at the higher level.

This case is hit in GraphicsComposerAidlCommandTest#PresentDisplay.

Change-Id: I03e5accc3ccda18426908d7d163f23d81c3145ea
Signed-off-by: Drew Davenport <ddavenport@google.com>