drm_hwcomposer: Fix cursor-related composition fallback
This change fixes an issue during display validation where
cursor layer failures would trigger a fallback that forces
all other layers to be client composited.

Prior to this fix, if the cursor layer could not match to
the cursor plane, DrmKmsPlan would attempt to match it to
another plane. This could result in the number of layers
exceeding the number of available planes, which causes the
composition to fail into the client-composition fallback
behavior.

This fix changes the first-level fallback behavior, so
that when the cursor layer doesn't match to the cursor
plane, the client range will be recaluclated in order to
reserve an additional plane for device composition. Then
the cursor layer is converted to device composition and
the composition is retried.

Change-Id: Id704e182b7cc1dc66270de45a986a29e4b71a5a8
Signed-off-by: Andrew Wolfers <aswolfers@google.com>