drm_hwcomposer: Add cursor stats
This change adds stats for commits that use the cursor plane.
The failed_kms_cursor_validate_ stat is incremented whenever a
test commit fails which attempted to match a layer to the cursor
plane. This event was previously being tracked by the
broader failed_kms_validate_ stat, but this change disentangles
the two events to help with debugging.

This change also adds a stat which tracks the number of frames
which successfully commit to the cursor plane.

Change-Id: I3f23e3e1d4bfd5ac70d2ceb1d768a2ded8ee034b
Signed-off-by: Andrew Wolfers <aswolfers@google.com>