drm_hwcomposer: Remove ChosePreferredConfig
This is only called from Init, so move the implementation to there.
Additionally, move it to the end of Init, which is fine since there is
no interdependency between ChosePreferredConfig and the rest of the Init
function.

With the logic moved to the end of the function, refactor the
conditionals to simplify, while ensuring that SetActiveConfig doesn't
get called when updating the configs fails.

Change-Id: I2cb5964f562413df70005666529ae2c820a1e5f9
Signed-off-by: Drew Davenport <ddavenport@google.com>