drm_hwcomposer: CI: Build aospless arm64 tarball
Currently the aospless arm64 tarball is built manually outside of CI,
uploaded to the project, and then downloaded when the CI runs.

Instead, replace this manual process with a script that builds the
aospless arm64 tarball and saves the tarball inside the aospless container
for use later in the pipeline.

This ensures that the method of building the tarball is documented and
repeatable.

Signed-off-by: Deborah Brouwer <deborah.brouwer@collabora.com>