drm_hwcomposer: CI: Add a cuttlefish container
Add a container to build cuttlefish images. The container makes the
compressed cuttlefish images available to subsequent jobs that need to
emulate an x86_64 phone running android.

The cuttlefish image is customized to use mesa llvmpipe and lavapipe
drivers.

The container build takes about 5 hours so it is expected that this
would be run very rarely.

Signed-off-by: Deborah Brouwer <deborah.brouwer@collabora.com>