drm_hwcomposer: CI: Add job to make cuttlefish artifacts
Add a new job `use-debian/x86_64_cuttlefish` to extract the compressed
cuttlefish image from the cuttlefish container and make it available as
a short-lived artifact for use by subsequent jobs in the pipeline.

The cuttlefish images need to be accessed as artifacts because the
cuttlefish container itself does not include all of the android tools
and tests needed to run cts and vts jobs.

Signed-off-by: Deborah Brouwer <deborah.brouwer@collabora.com>