drm_hwcomposer: CI: use dev fork to build container
If a contributor is a member of an upstream project, then the built-in
GitLab variable CI_PROJECT_URL points to the upstream project. This
means that a developer's fork in a merge request can't be used to build
a container.

Replace CI_PROJECT_URL and point to the developer's fork repository
instead.

Signed-off-by: Deborah Brouwer <deborah.brouwer@collabora.com>