drm_hwcomposer: Remove error handling from PresentStagedComposition
BadLayer could be returned in a few cases:
  1. Client Layer is not usable for device composition
  2. Some layer tagged for DEVICE composition is not usable for device
     composition
  3. Writeback layer is not usable for device composition

In the Validate phase, these errors are handled by falling back to GPU
composition. In the present phase, these are real errors.

The check was supposed to gracefully handle a case where the Client does
a validate/present cycle without having created any layers. This case is
handled by checking for empty layers and returning early, as described
in
https://gitlab.freedesktop.org/drm-hwcomposer/drm-hwcomposer/-/merge_requests/354

Any other as-yet-unknown cases where there are no layers to present should not
fail silently and raise an error.

Change-Id: Ie5fc2202936209a80ca8af5c3325520a1897b2e5
Signed-off-by: Drew Davenport <ddavenport@google.com>