drm_hwcomposer: Add thread annotations for DrmAtomicStateManager
Shared state is protected by a combination of the local mutex_ and the
global lock. This change attempts to document and enforce that more
clearly using clang-tidy thread annotations.

Change unique_lock usage to lock_guard for cases that do not require
release of the mutex such as when using a condition variable.

Temporarily add a main_lock_ mutex which mirrors how
the global lock is being used for synchronization. Lock this at the
entry point to all public methods. This lock will be removed as a part
of subsequent refactoring and it used to more cleanly break up the
following changes in this series.

Some state is not properly protected by any lock, so temporarily disable
the warnings around access to that state until the synchronization is
resolved in subsequent changes.

Change-Id: I6f4f30c4e97f266146b0879c5255d6f2432f12a3