ANDROID: drm_hwcomposer: Workaround for xe driver
The kernel uprev and/or transition from i915 to xe seems to have
uncovered a bug that results in significant graphical artifacts.

This can be mitigated by ensuring that already-signaled fences are not
committed for DEVICE layers that are not getting updated buffers, which
results in the driver using implicit fencing for the buffers.

Work around the issue to give some breathing room while the root cause
is investigated.

Bug: b:398935643
Flag: Excempt: desktop only
Test: No graphical artifacts with shadertoy on brya
Ignore-Upstream-First: Temporary downstream workaround
Change-Id: Idddf0828bc486fa6d110b3a91ff0a33afe2f4d43