drm_hwcomposer: Use most recently committed state
When preparing the properties for a commit, the state for the currently
active frame is used, rather than the last committed frame. This is
incorrect and could lead to a scenario such as:

- Frame N: commit enables plane 1
- Frame N+1: commit uses plane 1, 2, 3
- Frame N+2: if executed before staged_frame_state_ is moved to
    active_frame_state_, the new KmsState::used_framebuffers will be
    initialized to {1}
  - commit uses planes 1, 2
  - plane 3 is not disabled by the unused_planes code in CommitFrame

A similar problem exists for the KmsState::crtc_active_state.

To avoid this issue, always use the last committed state, regardless of
whether the previous frame has been presented or not.

Change-Id: I840b5b0ca3a3cf477882cacf6b1d5c0313ae80cf