drm_hwcomposer: Update DrmKmsPlan to position cursor plane with highest z-pos
Cursor composition was previously developed under the mistaken assumption
that the z-pos property of the cursor plane was immutable. The DrmKmsPlan
would always place the cursor plane at the lowest z-position, but the
immutability made that aspect moot.

This change updates plan creation to instead place the cursor plane at the
highest z-position, which fixes cursor positioning on hardware where the
cursor plane z-pos is mutable.

Change-Id: Idfd390e0eb3de8ff6125722cda9f5ce73e983a02
Signed-off-by: Andrew Wolfers <aswolfers@google.com>