drm_hwcomposer: Check for IStats service as needed
IStats service is started by system_server which is not running during
VTS tests. In this case, the IStats service is declared, but not
started, which may result in a hang when calling AServiceManager_waitForService.

Acquire the IStats client lazily when reporting atoms, and log an error
upon failure to do so. If IStats is declared, we should expect that the
service will be available under normal operation.

Another benefit of this approach is that acquiring the IStats service is
no longer blocking service initialization.

Change-Id: Ib7a38930b1a3cbcaa629e576f601d7a05980e257