ANDROID: drm_hwcomposer: New workaround for xe driver
Since the issue appears to be related to handling of fences in the
kernel, sync on the fences in userspace before committing.

This seems to result in fewer graphical artifacts than the previous
workaround.

The workaround can be disabled for debugging purposes by setting
vendor.hwc.drm.enable_xe_workaround to 0

Bug: b:398935643
Flag: Exempt: desktop only
Test: No graphical artifacts with shadertoy on brya
Ignore-Upstream-First: Temporary downstream workaround
Change-Id: Ia2f053cef1c909544ed6f92c86f1fd00304f9c83