drm_hwcomposer: Do not clear ctm after commit.
Keep reference to ctm so that we can switch between identity CTM and
requested CTM. When the client composites all layers, all color
transforms are also applied by the client. In these cases we apply an
identity CTM. When we swap to device composition, we need to re-apply
the CTM.

Additional code cleanup.

Change-Id: I8862b37ad6f5f785aba78f9172b321ad159c2e89