drm_hwcomposer: CI: Save CTS failure logs
When $FDO_CI_BASH_HELPERS was added, it included an ERR trap that is
triggered when a cts failure is found in the invocation summary. This
prevents the full cts logs from being saved in the job results.

Revise the cts failure check to avoid triggering the trap on error and
allow the full cts logs to be saved.

Signed-off-by: Deborah Brouwer <deborah.brouwer@collabora.com>