drm_hwcomposer: Calcuate better estimate for refresh_time_ns
QueueConfig allows for a seamless refresh rate change to be queued at
some arbitrary point in the future.

Previous behavior was to report `desired_time` as the time to request
a new frame from the client to apply the refresh rate change. Since
`desired_time` could be any arbitrary point in time, this may not align
with the current refresh rate.

This change addresses a failure in
SetActiveConfigWithConstraints_Delayed exposed by the previous change to
avoid setting the active_config_id until a full commit.

Change-Id: I92888cd1a6a330455f9e7e496d6900d494605248