drm_hwcomposer: Return kUnsupported for setLayerLuts
Even though drmhwc does not support per-layer LUTs, the vts test
GraphicsComposerAidlCommandV4Test.SetUnsupportedLayerLuts issues the
setLayerLuts command when the display is in headless mode with
vkms/cuttlefish. Since drmhwc ignores the command, the test fails.

Instead of ignoring the command, return a kUnsupported error which
allows the test to pass.

Signed-off-by: Deborah Brouwer <deborah.brouwer@collabora.com>