drm_hwcomposer: CI: Update vts tests to skip
The current list of tests to skip from VtsHalGraphicsComposer3_TargetTest
is from Android15, but the CI is running Android16, so update the list
of tests that are expected to skip from Android16.

This make the CI more accurate, because if a change causes a new test to
skip unexpectedly, the CI job will fail.

Signed-off-by: Deborah Brouwer <deborah.brouwer@collabora.com>