drm_hwcomposer: CI: launch_cvd: run modem_simulator
The modem simulator is currently disabled because, in an earlier version
of cuttlefish, it was preventing drmhwc from launching. However, the
missing service causes a lot of noise in logcat as the ServiceManager
continuously looks for the IRadioModem service.

Since the cuttlefish modem simulator is now working with drmhwc on
Android16, enable the modem simulator and reduce the logcat and Activity
Manager noise.

Also, while here, remove the executable bit from launch-cvd.sh since it
is only sourced, not run directly.

Signed-off-by: Deborah Brouwer <deborah.brouwer@collabora.com>