drm_hwcomposer: CI: Add vts job for multi-displays
Currently VtsHalGraphicsComposer3_TargetTest is run only with a single
display. Use the configurable vkms driver to run vts with multiple
displays.

This job should help to catch errors that arise only when there are
multiple displays. It also allows the test
ConnectedDisplaysTest.IndependentConfigChange to run whereas it is
otherwise skipped for single display configurations.

Signed-off-by: Deborah Brouwer <deborah.brouwer@collabora.com>