drm_hwcomposer: Do not set color_matrix_ for 3x4 transform
When the request CTM is a 3x4 transform, we designate all layers for
client composition and do not set the plane CTM. In the DRM_OR_IGNORE
case, we do not force client composition and end up using the incomplete
transform.

This change ensures that we only set the CTM when the matrix is useable.

Change-Id: Ic7928616208b61fcdef72cc214d652cdf8a47d00