CI: Set a cache directory for mesa
If a shader cache directory is not specified, mesa attempts to make a
cache directory at root. But since cuttlefish root is read-only, this
creates errors which fail the CI jobs.

Specify a shader cache directory for mesa to use that is writable and
will not cause spurious errors in CI.

Signed-off-by: Deborah Brouwer <deborah.brouwer@collabora.com>