Restore cached DrmKmsPlan reuse
This change restores behavior where the DrmKmsPlan cached during
ValidateStagedComposition is reused during Present. That behavior
was disabled in a prior change due to a bug where the client
layer could become stale if it were updated between the validate
and present steps, which is now resolved by this change.

Change-Id: Ibca43dc028ece2f4480e5631078ad51320c10582