Populate LayerData as a part of SetLayerProperties
There is no scenario where we'd not want to populate the LayerData after
setting the layer properties.

Previous behavior was to call PopulateLayerData for layers with DEVICE
and CURSOR type, as well as special cases like the client target layer
and writeback layer. In all of these cases, active_slot_id_ will be set.
Whenever active_slot_id_ is set, call PopulateLayerData.

Change-Id: I0d92e8e1f613bf8456defc324aedbf259b895a36