Remove hwc2_device
These files are not being built or used, and their presence has been a
consistent source of confusion for contributors to drmhwc.

One of the rationales for keeping these around was to help enforce the
separation between the HWC3 "frontend" and the drmhwc midlayers, but
the unit tests which were added since then help maintain this structure
and seperation.

Change-Id: Iec1656765f2839113e51ef2330d23dc9bdb5cfb5