Remove duplicate GrallocBufferHandle caching
GrallocBufferHandle needs to be kept alive for the duration of its
associated BufferInfo. This is accomplished by holding
std::shared_ptr<GrallocBufferHandle> on the BufferInfo itself, as well
as in a map in the Hwc3Layer class.

Since the BufferInfo in the HwcBufferCache holds a reference to the
shared_ptr, it's not necessary to also hold a reference to it in a
separate std::map.

Drop the additional std::map and rely on the HwcBufferCache to ensure
that there is at least one shared_ptr<GrallocBufferHandle> keeping the
buffer alive until the buffer slot is cleared.

Change-Id: Ic1d3fad3a402aea53a829ce8897698225260b98b