Don't cache Hwc3BufferHandle on error
If we successfully import the Hwc3BufferHandle, but fail to extract a
BufferInfo from it, don't cache the imported Hwc3BufferHandle.

Subsequent fetches from the cache would fail anyways, and there is no
benefit from caching the imported handle since there are no resources
(i.e. fds) that need to be kept open.

Change-Id: I3c15e9f723f8b72b065695b8a06d49f94d7f5809