Generate 1D LUT for color pipeline
Use the guideline in Dataspace.aidl to create 1D LUTs for tonemapping
pixel data. The LUTs are generated for per display and cached to be used
among multiple layers.

The value 750 / 10000 is used as a temporary placeholder for HDR
headroom. This was tested locally to be a passable value in the general
use case.

Change-Id: Ib252ae540c19ce7d31156c7ca3b562ab8ae98085